pyproject.toml
[tool.poetry]
name = "backpack-tf2"
version = "0.1.0"
description = ""
authors = ["Danila Koshevoi <dmkosh10@mts.ru>"]
readme = "README.md"
packages = [{include = "backpack_tf2"}]

[tool.poetry.dependencies]
python = "^3.11"
oauthlib = "^3.2.2"
flake8 = "^6.0.0"
black = "^23.1.0"
python-dotenv = "^0.21.1"
backpacktf = "^0.1.4"
requests-oauthlib = "^1.3.1"
pydantic = "^1.10.5"
pymongo = "^4.3.3"
discord-webhook = "^1.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
